refactor: 抽取明细表冻结行列逻辑，交叉表冻结首行逻辑复用 (#2415)
* refactor: extract freeze public logic

* fix: lint error fixed

* fix: follow variable naming conventions

---------
